Task 006

+++ src/main/java/acme/features/anonymous/task/AnonymousTaskListService.java
@@ -42,7 +42,7 @@ public class AnonymousTaskListService implements AbstractListService<Anonymous, 
 		
 		Collection <Task>  result;
 		
-		result = this.repository.findTask().stream().sorted(Comparator.comparingInt(Task::getWorkload)).collect(Collectors.toList());
+		result = this.repository.findTask().stream().sorted(Comparator.comparingDouble(Task::getWorkload)).collect(Collectors.toList());
 
 		return result;
 	}
